requirements.txt
Babel==1.3
blinker==1.3
coverage==3.7.1
decorator==3.4.2
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-OpenID==1.2.4
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.11
flipflop==1.0
guess-language==0.2
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pbr==0.11.0
python-openid==2.2.5
pytz==2015.4
six==1.9.0
speaklater==1.3
SQLAlchemy==0.9.9
sqlalchemy-migrate==0.9.6
sqlparse==0.1.15
Tempita==0.5.2
Werkzeug==0.10.4
wheel==0.24.0
Whoosh==2.7.0
WTForms==2.0.2
urllib3==1.2.2
beautifulsoup4==4.4.0
requests==2.9.1
trueskill==0.4.4
iso8601==0.1.11
psycopg2==2.6.1
